Generators/NewGeneratorReturn: add extra test
Since PHP 8.3 comments are allowed between the `yield` and `from` keywords in a `yield from` expression.
As of PHPCS 3.11.0, this will be properly supported by PHPCS itself (PR to add support is currently open).

This commit only adds an extra test to the `NewGeneratorReturn` sniff to safeguard that this will not cause the sniff to throw duplicate error messages.

Related to:
* php/php-src 14926
* PHPCSStandards/PHP_CodeSniffer 529
* PHPCSStandards/PHP_CodeSniffer 647