Keywords/NewKeywords: handle "yield from" PHPCS cross-version
Follow up on 1478

The PHP 8.3 tokenization change for `yield from` with a comment between will be handled more consistently (and in the same way PHP cross-version) as of PHPCS 3.11.0.

This updates the sniff to handle this correctly for both "PHP 8.3+ with PHPCS < 3.11.0" as well as "PHPCS 3.11.0", while still preventing duplicate messages for multi-token `yield from` keywords.

Includes an additional test to safeguard this.

Related to:
* PHPCSStandards/PHP_CodeSniffer 529
* PHPCSStandards/PHP_CodeSniffer 647