PHP 8.3 | :sparkles: New `PHPCompatibility.Generators.NewYieldFromComment` sniff
> . In addition to whitespace characters, now comments are allowed between
>   `yield` and `from`. The whole "construct" (e.g. `yield /* comment */ from`)
>   is reported as a single `T_YIELD_FROM` token by the tokenizer.

This commit adds a new sniff which detects this in a PHPCS cross-version compatible manner.

Once support for PHPCS < 3.11.0 has been dropped, this sniff can be simplified, but IMO, this sniff alone is not enough reason to raise the minimum supported PHPCS version, especially as working around the different tokenizations isn't that complicated.

Includes tests.
Includes documentation.

Refs:
* php/php-src 14926
* https://github.com/php/php-src/blob/e4010fcf957c62ddb70f8d3296d52c5486a84d4d/UPGRADING#L48-L50
* php/php-src 10125
* https://github.com/php/php-src/commit/f291d37a1a7d78e841f40a4410359548bc73de1b

Related to 1589
Fixes 1720